flask_app/requirements.txt
email_validator==2.2.0
Flask==3.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.0.0
nodezator==1.5.1
psycopg2-binary==2.9.10
Prefect==3.0.10
python-dotenv==1.0.1
PyYAML==6.0.2
Werkzeug==3.0.4
WTForms==3.0.1